Attempt a smarter undistortion, with cropping.
As an exploration, I did some undistortion on a video from a gopro
hero+, which comes with a pretty wide-angle lens. Code at
https://goo.gl/YtHbFJ , before at https://goo.gl/S9tWyR , after at
https://goo.gl/zDcozC . The checkerboard is undistorted, yes, but there
are strong new distortions introduced at the edges. Is there a way to
deal with this other than just cropping? Also, I couldn't use all 1951
of my recorded frames for calculating the undistortion, or the call to
cv2.calibrateCamera would be the way too slow. What's a good rule of
thumb for how many I should actually use? 100 seemed to be about the
upper limit for reasonable runtime.

NOVEMBER 26, 2017
Hello Tom, have you started or completed Advanced Lane Finding lesson
and project? There you will find a lot of tools and ways on working
around distortion and correcting it. Regarding your question about
calibrating the camera, for project for we use really little amount of
images to calibrate the camera.
have you seen these pages?
http://opencv-python-
tutroals.readthedocs.io/en/latest/py_tutorials/py_calib3d/py_calibration/py_calibration.html
http://answers.opencv.org/question/60786/how-many-images-for-camera-
calibration/
Please let me know if that helps
4:20 am

Image
Huh. I updated the video links. It seems that I can use
getOptimalNewCameraMatrix to get a second camera matrix and a region-of-
interest, and this sometimes works. In fact, when it works, I don't even
need the RoI--turning the alpha parameter on this function down from 1
to say .75 does some cropping of its own. The cropping imposed by the
RoI is extreme--it basically negates the original wide-angle feature of
the gopro camera I'm using.
Image
However, when it doesn't work, it fails pretty hard. I intuit that the
second camera matrix, provided by getOptimalNewCameraMatrix for the
image being undistorted, somehow very different from the one found by
calibrateCamera. This seems to happen only some of the time (I shuffle
images before finding their object points for calibrateCamera), so maybe
I could check the difference between these two camera matrices, and, if
they say have a large Frobenius-norm difference, reshuffle the images
and try again.
But this only seems to be a problem for the severely-fisheye'd images
from my gopro, not for the more mildly distorted images for the
assignment, so I'll table this for now.