Remove half-baked parallel type system.
I still think there might be a "right" way to do this (which may or may
not rely on Python's type system), but I'm not going to do it now. It's
a solution seeking a problem--there are certainly others in this
project, and I'll need to stop with that behavior now.

I Shall Exterminate Everything Around Me That Restricts Me from Being
the Master