Fix maxFrames argument.
Wow, this new plotter is 14 times faster. Matplotlib really should not
be used for video (at least, not in the way I use it. I'm unwilling to
use the manimation approach and alter the data of pre-plotted lines;
that is just too much work, and it's real use cases are always
harder/impossibler than their examples make it seem.)